Recursive tuple and lists support
Implementing changes by @sdh4, issue #24

This passes all tests, but stores simple lists in a non-ideal way.

A new test to differentiate long and int types in a list has been added; currently this test is failing.

This may become the first modification towards pickle v2.0, which will require adding a version attribute so that v1 and v2 pickles
can be differentiated.