Adding failing tests to capture issues:
https://github.com/telegraphic/hickle/issues/24
https://github.com/telegraphic/hickle/issues/26

These issues need to be fixed asap. It may however be prudent to make refactor
changes first, in prep for v2 hickle.

At the root of the issues is how to store arbitrary iterable data. For example,
a python list can contain only integers, or it can contain heterogeneous items.

What is the best method for storing these, without significant speed penalties?
And how does one make sure that when unpickling, the data order remains the same?