Iterative dump and load now works for sets, lists, and tuples
Man, that was way more subtle than anticipated. Need to be careful
as tuple and set methods do not have an append() function, which is
crucial for iterative spidering of arbitrary nested list,tuple,sets.

Next up is to enforce that order is preserved, by sorting group IDs.
One concern of complete arbitrariness will be speed: there may be
a few copies of large datasets required. For now, getting it working
is the most important part.