Added scipy sparse matrix support
So far only CSC and CSR matrices. Brings up a question of how best to
provide "plugin" support for desired functionality. For example, how
to add support for astropy classes, and other packages with weird and
wonderful objects in them? The code will get pretty messy if a plugin
system isn't setup.