Refactored code to add loaders/ directory
Idea here is to make it straightforward for users to add their own
loaders and dumpers. In particular, I would like to add support for
some astropy and pandas types. Packages such as these won't be needed
by every user, so separating out dump/load into self-contained files
will make it easier to ensure these are optional.

A directory loaders/ has been added, with the intention of this becoming
a plugin-style directory, into which new loaders can be added by users as
required. The next step is to tackle `create_dataset_lookup` so that the
dump/load methods can be added to the lookup dictionary easily.