First very basic python3 support!
Finally getting toward py3 support (issue #8):
https://github.com/telegraphic/hickle/issues/8

So far, the only test that passes is dumping/loading a list.
Still, a big milestone!

The first step will be to rewrite all the loaders in load_python.py
and create equivalent versions in load_python3.py.

There is a lot of meddling still to do, but I think the general
hard work in refactoring to make this possible is done. The bit
that will be fiddly is dealing with bytearray/bytes/strings as
loaded in and out of h5py in python3. This will take some getting used to.