Proper handling of file and file like objects
Hickle 4.0.x accepts file like objects and uses them as model for
creating a new file discarding file handler passed in. If the file
should be accessed later on using the same file handler this results
in an IOError raised and alike.

The proposed fix does not close the file but checks that it is opened at
least read able if mode is 'r' or read and writable in case mode reads
'w', 'r+', 'w+','x', 'x+' or 'a'.

Note: passed in file or file like object has to be closed by caller.